Add files via upload
TRIBUTE PAGE :

A tribute page built with HTML and CSS. 
(https://github.com/snehabarik/oibsip_taskno).

 Description:

This repository contains a tribute page developed as part of the OIBSIP task. The page is designed to honor and celebrate the life and achievements of a notable individual. Built with HTML and CSS, it features a clean and responsive design, ensuring an optimal viewing experience across different devices. 
(https://github.com/snehabarik/oibsip_taskno).

 Features:

- Clean and modern design
- Responsive layout
- Information about the individual's life and achievements

 Installation:

1. Clone the repository:
   ```bash
   git clone  https://github.com/snehabarik/oibsip_taskno.git